requirements.txt
accelerate==1.7.0
agno==1.5.4
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.16.1
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1743190139881/work
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
api==0.0.7
archspec @ file:///croot/archspec_1709217642129/work
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boltons @ file:///croot/boltons_1737061692168/work
Brotli @ file:///croot/brotli-split_1736182456865/work
build==1.2.2.post1
cachetools==5.5.2
certifi @ file:///croot/certifi_1745939216646/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer==3.4.2
chromadb==1.0.10
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
conda @ file:///croot/conda_1743715288469/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1744662537425/work
conda-anaconda-tos @ file:///croot/conda-anaconda-tos_1744823864153/work
conda-content-trust @ file:///work/perseverance-python-buildout/croot/conda-content-trust_1728487058234/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1745607008911/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1731366181659/work
coverage==7.8.0
cryptography @ file:///croot/cryptography_1740577825284/work
Deprecated==1.2.18
dill==0.4.0
distro @ file:///work/perseverance-python-buildout/croot/distro_1728396110052/work
dnspython==2.7.0
docling==2.34.0
docling-core==2.31.2
docling-ibm-models==3.4.3
docling-parse==4.0.1
docstring_parser==0.16
durationpy==0.10
easyocr==1.7.2
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
faiss-cpu==1.11.0
fastapi==0.115.9
fastapi-cb==1.1.0
fastapi-cli==0.0.7
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
Flask-Limiter==3.12
flatbuffers==25.2.10
frozendict @ file:///work/perseverance-python-buildout/croot/frozendict_1728497542215/work
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.94.0
google-cloud-bigquery==3.33.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.16.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gptcache==0.1.44
greenlet==3.2.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonlines==3.1.0
jsonpatch @ file:///work/perseverance-python-buildout/croot/jsonpatch_1728399595941/work
jsonpointer==2.1
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
langchain-core==0.3.59
langgraph==0.4.3
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.42
latex2mathml==3.78.0
lazy_loader==0.4
libmambapy @ file:///croot/mamba-split_1734469461757/work/libmambapy
limits==5.1.0
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
marko==2.1.3
MarkupSafe==3.0.2
mdurl==0.1.2
menuinst @ file:///croot/menuinst_1738943416351/work
mmh3==5.1.0
motor==3.7.1
mpire==2.10.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.18
networkx==3.4.2
ninja==1.11.1.4
nose==1.3.7
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
ollama==0.4.8
onnxruntime==1.22.0
openai==1.78.1
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
ordered-set==4.1.0
orjson==3.10.18
ormsgpack==1.9.1
overrides==7.7.0
packaging==25.0
pandas==2.2.3
passlib==1.7.4
peft==0.15.2
pgvector==0.4.1
pillow==11.2.1
platformdirs @ file:///croot/platformdirs_1744273042065/work
pluggy @ file:///croot/pluggy_1733169602837/work
posthog==4.2.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycosat @ file:///croot/pycosat_1736868416091/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.11.4
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments @ file:///croot/pygments_1744664109463/work
pylatexenc==2.10
pymongo==4.13.0
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1728386193338/work
pytest==8.3.5
pytest-asyncio==1.0.0
pytest-cov==6.1.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.2
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich @ file:///croot/rich_1732638981168/work
rich-toolkit==0.14.6
rpds-py==0.25.1
rsa==4.9.1
rtree==1.4.0
ruamel.yaml @ file:///croot/ruamel.yaml_1745960305322/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1745937152469/work
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
semchunk==2.2.2
sentence-transformers==4.1.0
setuptools==80.7.1
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlglot==26.21.0
starlette==0.45.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.5.24
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.0
torchvision==0.22.0
tqdm @ file:///croot/tqdm_1738943501192/work
transformers==4.51.3
triton==3.3.0
truststore @ file:///croot/truststore_1736550121485/work
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.4.0
uv==0.7.7
uvicorn @ file:///croot/uvicorn-split_1732660726596/work
uvloop==0.21.0
waitress==3.0.2
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard @ file:///croot/zstandard_1731356346222/work
